Lecture 63: Finish loggings

+++ src/tasks/tasks.service.ts
@@ -20,9 +20,9 @@ export class TasksService {
     return this.taskRepository.getTasks(filterDto, user);
   }
 
-  async getTaskById(id: number): Promise<Task | undefined> {
+  async getTaskById(id: number, user: User): Promise<Task | undefined> {
     const found = await this.taskRepository.findOne({
-      where: { id },
+      where: { id, userId: user.id },
     });
     if (!found) {
       throw new NotFoundException(`Task with ID ${id} not found`);
@@ -34,24 +34,22 @@ export class TasksService {
     return this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async deleteTaskById(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTaskById(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID ${id} not found`);
     }
     console.log(`Deleted task with ID ${id}`);
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
   }
-
-  // updateTask(id: string, status: TaskStatus): Task {
-  //   const getTask = this.getTaskById(id);
-  //   getTask.status = status;
-  //   return getTask;
-  // }
 }
